+++ NEWS.md
@@ -1,7 +1,7 @@
 # ggtree 2.1.1
 
 + convert roxygen documents using markdown (2019-11-01, Fri)
-+ `xlim_align` and `ylim_align` to uniformize axis limits of ggplot objects (2019-10-31, Thu)
++ `xlim2` and `ylim2` to uniformize axis limits of ggplot objects (2019-10-31, Thu)
   - <https://yulab-smu.github.io/treedata-book/chapter7.html#composite_plot>
   - <https://yulab-smu.github.io/treedata-book/docs/chapter10.html#axis_align>
 + fixed `fontface` warning message when `align=TRUE` in `geom_tiplab` (2019-10-30, Wed)